refactor: Refactor ITypeParameterRepository import paths
