+++ src/application/use-cases/typeParameter/DeleteTypeParameterUsecase.ts
@@ -1,4 +1,4 @@
-import { ITypeParameterRepository } from "../../../domain/models/entities/TypeParameter";
+import { ITypeParameterRepository } from "../../../domain/interfaces/repositories/ITypeParameterRepository";
 
 export default class DeleteTypeParameterUseCase {
     constructor(private typeParameterRepository: ITypeParameterRepository) {}
