+++ src/application/use-cases/typeParameter/ListTypeParameterUsecase.ts
@@ -1,4 +1,4 @@
-import { ITypeParameterRepository } from "../../../domain/models/entities/TypeParameter";
+import { ITypeParameterRepository } from "../../../domain/interfaces/repositories/ITypeParameterRepository";
 
 export class ListTypeParameterUseCase {
     constructor(private typeParameterRepository: ITypeParameterRepository) {}
