+++ src/application/use-cases/typeParameter/ReadTypeParameterUseCase.ts
@@ -1,4 +1,4 @@
-import { ITypeParameterRepository } from "../../../domain/models/entities/TypeParameter";
+import { ITypeParameterRepository } from "../../../domain/interfaces/repositories/ITypeParameterRepository";
 import { TypeParameter } from "../../../domain/models/entities/TypeParameter";
 
 export class ReadTypeParameterUseCase {
