+++ src/application/use-cases/typeParameter/UpdateTypeParameterUseCase.ts
@@ -1,6 +1,6 @@
-import { ITypeParameterRepository } from "../../../domain/models/entities/TypeParameter";
+import { ITypeParameterRepository } from "../../../domain/interfaces/repositories/ITypeParameterRepository";
 import { TypeParameter } from "../../../domain/models/entities/TypeParameter";
-import UpdateTypeParameterDTO from "../../../web/dtos/typeParameters/UpdateTypeParameterDTO"; 
+import UpdateTypeParameterDTO from "../../../web/dtos/typeParameter/UpdateTypeParameterDTO"; 
 
 export default class UpdateTypeParameterUseCase {
     constructor(private typeParameterRepository: ITypeParameterRepository) {}
