+++ src/domain/interfaces/repositories/ITypeParameterRepository.ts
@@ -1,0 +1,10 @@
+import { TypeParameter } from "../../models/entities/TypeParameter";
+
+export interface ITypeParameterRepository {
+    create(typeParameter: Partial<TypeParameter>): Promise<TypeParameter>;
+    delete(id: string): Promise<boolean>;
+    update(id: string, typeParameter: Partial<TypeParameter>): Promise<TypeParameter | null>;
+    list(): Promise<TypeParameter[]>;
+    findById(id: string): Promise<TypeParameter | null>;
+    findByName(name: string): Promise<TypeParameter | null>;
+}
